remake collision detection

+++ mainwindow.h
@@ -53,7 +53,7 @@ private:
     int dispWidth;
     int dispHeight;
     QPoint dispCenter;
-    int dispPadding = 150;
+    double dispPadding = 150;
 
 
     QDot* red_dot;
